build(cmake): make the game 64-bit and drop the 32-bit build
x64 is the only build now: `-A Win32` is a configure error, and every
`#if defined(_M_IX86)` is gone from the tree rather than guarded.

Sound and video no longer bind to the game's own DLLs, which are 32-bit
images a 64-bit process cannot load at all. Miles6/xaudio2 implements the
Miles surface over XAudio2, with FFmpeg for anything streamed and an IMA
ADPCM decoder for the game's compressed samples; Bink/ffmpeg implements
the Bink surface over FFmpeg, which decodes binkvideo and binkaudio
natively, so the movies on disk play as they are. MilesAudioManager.cpp
and BinkVideoPlayer.cpp are untouched and still call AIL_* and Bink*.

The assembly was the easy half. What cost the most was four places that
assumed a pointer fits an int, none of which a 32-bit build can show:
StringClass's temp-string pool, ObjectPoolClass's block header, the debug
library's dbghelp function table, and the GUI's WindowMsgData. One more
sat in a file format: LegacyDDSURFACEDESC2 held a void*, grew by four
bytes, and made every DDS in the game a zero-height texture, which
crashed the texture loader under a particle effect. RLE's compressor read
one byte past its buffer, which was wrong on both architectures.

Both crash-log stack walkers unwind from the exception tables now, and
the registry read asks the 32-bit view as well, because the installers
that write the game's path are 32-bit and a 64-bit process does not see
their keys.

Checked rather than assumed. Same seed, same generated map, 3000 frames
of AI against AI: HEADLESS CRC 0x676E1EFA from the old 32-bit build and
from this one. Replays play back to the same world. The picture is within
sub-pixel sampling of the 32-bit build over eight views, 0.03% of blocks
differing at quarter size. miles_smoke plays a tone as a sample and as a
stream and waits for the callbacks a voice only fires once it has played
the buffer out; bink_smoke decodes a real .bik against its own header.

+++ GeneralsMD/Code/Libraries/Source/debug/debug_stack.h
@@ -59,7 +59,8 @@ public:
     unsigned m_numAddr;
 
     /// addresses
-    unsigned m_addr[MAX_ADDR];
+    // Code addresses, so pointer sized rather than four bytes.
+    size_t m_addr[MAX_ADDR];
 
   public:
     explicit Signature(void): m_numAddr(0) {}
@@ -81,7 +82,7 @@ public:
       \param n index, 0..Size()-1
       \return signature address
     */
-    unsigned GetAddress(int n) const;
+    size_t GetAddress(int n) const;
 
     /**
       \brief Strong ordering operator.
@@ -113,7 +114,7 @@ public:
       \param buf return buffer
       \param bufSize size of return buffer, minimum is 64 bytes (256 recommended)
     */
-    static void GetSymbol(unsigned addr, char *buf, unsigned bufSize);
+    static void GetSymbol(size_t addr, char *buf, unsigned bufSize);
     
     /**
       \brief Determines symbol for given address.
@@ -130,7 +131,7 @@ public:
       \param line line number, may be NULL
       \param relLine relative address within line, may be NULL
     */
-    static void GetSymbol(unsigned addr,
+    static void GetSymbol(size_t addr,
                           char *bufMod, unsigned sizeMod, unsigned *relMod,
                           char *bufSym, unsigned sizeSym, unsigned *relSym,
                           char *bufFile, unsigned sizeFile, unsigned *line, unsigned *relLine);
